make harris make plots (#1027)
